+++ components/note/card/Reminder.vue
@@ -1,6 +1,6 @@
 <template>
   <v-card class="elevation-6">
-    <v-toolbar color="primary" flat>
+    <v-toolbar :color="titleColor" flat>
       <v-toolbar-title>
         <ClickToEdit :value="note.title" @input="onTitleChange" />
       </v-toolbar-title>
@@ -12,7 +12,7 @@
 
     <v-dialog v-model="fullscreen" fullscreen hide-overlay transition="dialog-bottom-transition">
       <v-card>
-        <v-toolbar color="primary" flat>
+        <v-toolbar :color="titleColor" flat>
           <v-toolbar-title>{{note.title}}</v-toolbar-title>
           <div class="flex-grow-1"></div>
           <v-btn icon @click="fullscreen = false">
@@ -25,7 +25,7 @@
             <v-row no-gutters justify="center">
               <v-col cols="11" align-self="center" class="text-right pr-3">{{formattedDate}}</v-col>
               <v-col cols="1">
-                <v-icon :color="titleColor">alarm</v-icon>
+                <v-icon :color="alarmColor">alarm</v-icon>
               </v-col>
             </v-row>
           </h2>
@@ -55,7 +55,7 @@
         <v-row no-gutters justify="center">
           <v-col cols="11" align-self="center" class="text-right pr-3">{{formattedDate}}</v-col>
           <v-col cols="1">
-            <v-icon :color="titleColor">alarm</v-icon>
+            <v-icon :color="alarmColor">alarm</v-icon>
           </v-col>
         </v-row>
       </h2>
@@ -139,12 +139,15 @@
         if(!this.noteSettings.fixed) return ""
         return `card-view-${this.noteSettings.size}`
       },
-      titleColor(){
+      alarmColor(){
         let diff = dateFN.differenceInMinutes(this.note.content.date, new Date())
 
         if(diff <= 0) return 'error'
         if(diff <= 5) return 'secondary'
         return 'primary'
+      },
+      titleColor(){
+        return this.note.color || 'primary'
       },
       formattedDate(){
         return dateFN.format(this.note.content.date, this.$t('common.datetime.format.datetime'))