+++ components/note/form/Picture.vue
@@ -2,8 +2,11 @@
   <v-form v-model="valid" @submit.prevent="onSubmit()" :id="formId" ref="form">
     <v-container>
       <v-row>
-        <v-col cols="12">
+        <v-col cols="9" sm="11">
           <v-text-field v-model="note.title" :label="$t('note.title')" :placeholder="$t('note.untitled')"></v-text-field>
+        </v-col>
+        <v-col cols="3" sm="1">
+          <ColorPicker v-model="note.color" />
         </v-col>
         <v-col cols="12">
           <v-text-field v-model="note.url" :label="$t('note.text.content')" :placeholder="$t('note.picture.url.placeholder')" :rules="ruleRequired" required autofocus></v-text-field>
@@ -19,9 +22,10 @@
 <script>
   import { mapGetters } from 'vuex';
   import TagPicker from "./TagPicker";
+  import ColorPicker from "./ColorPicker";
   export default {
     name: "NoteFormPicture",
-    components: {TagPicker},
+    components: {ColorPicker, TagPicker},
     props: {
       formId: {
         type: String,
@@ -44,11 +48,16 @@
         type: Array,
         required: false,
         default: () => []
+      },
+      initialColor: {
+        type: String,
+        required: false,
       }
     },
     data() {
       let note = {
         title: "",
+        color: this.initialColor,
         tags: this.initialTags,
         url: "",
       }
@@ -56,6 +65,7 @@
       if(this.data) {
         note.title = this.data.title === this.$t('note.untitled') ? null : this.data.title
         note.tags = this.data.tags
+        note.color = this.data.color
         note.url = this.data.content.url
       }
 
@@ -92,6 +102,7 @@
         }
         data.title = this.note.title ? this.note.title : this.$t('note.untitled')
         data.tags = this.note.tags
+        data.color = this.note.color
         data.content.url = this.note.url
 
         this.$emit('onSubmit', data)
@@ -105,6 +116,7 @@
       data(newData) {
         this.note.title = newData.title === this.$t('note.untitled') ? null : newData.title
         this.note.tags = newData.tags
+        this.note.color = newData.color
         this.note.url = newData.content.url
       }
     }