+++ components/note/form/Record.vue
@@ -2,8 +2,11 @@
   <v-form v-model="valid" @submit.prevent="onSubmit()" :id="formId" ref="form">
     <v-container>
       <v-row>
-        <v-col cols="12">
+        <v-col cols="9" sm="11">
           <v-text-field v-model="note.title" :label="$t('note.title')" :placeholder="$t('note.untitled')"></v-text-field>
+        </v-col>
+        <v-col cols="3" sm="1">
+          <ColorPicker v-model="note.color" />
         </v-col>
       </v-row>
 
@@ -41,10 +44,12 @@
 <script>
   import { mapGetters } from 'vuex';
   import TagPicker from "./TagPicker";
+  import ColorPicker from "./ColorPicker";
 
   export default {
     name: "NoteFormRecord",
     components: {
+      ColorPicker,
       TagPicker,
     },
     props: {
@@ -69,11 +74,16 @@
         type: Array,
         required: false,
         default: () => []
+      },
+      initialColor: {
+        type: String,
+        required: false,
       }
     },
     data() {
       let note = {
         title: "",
+        color: this.initialColor,
         tags: this.initialTags,
         record: null,
         description: ""
@@ -82,6 +92,7 @@
       if(this.data) {
         note.title = this.data.title === this.$t('note.untitled') ? null : this.data.title
         note.tags = this.data.tags
+        note.color = this.data.color
         note.description = this.data.content.description
         note.record = this.data.content.record
       }
@@ -125,6 +136,7 @@
         }
         data.title = this.note.title ? this.note.title : this.$t('note.untitled')
         data.tags = this.note.tags
+        data.color = this.note.color
         data.content.description = this.note.description
         data.content.record = this.note.record
 
@@ -140,6 +152,7 @@
       data(newData) {
         this.note.title = newData.title === this.$t('note.untitled') ? null : newData.title
         this.note.tags = newData.tags
+        this.note.color = newData.color
         this.note.description = newData.content.description
         this.note.record = newData.content.record
       }