+++ components/note/form/Text.vue
@@ -2,8 +2,11 @@
   <v-form v-model="valid" @submit.prevent="onSubmit()" :id="formId" ref="form">
     <v-container>
       <v-row>
-        <v-col cols="12">
+        <v-col cols="9" sm="11">
           <v-text-field v-model="note.title" :label="$t('note.title')" :placeholder="$t('note.untitled')"></v-text-field>
+        </v-col>
+        <v-col cols="3" sm="1">
+          <ColorPicker v-model="note.color" />
         </v-col>
         <v-col cols="12">
           <v-textarea v-model="note.content" :label="$t('note.text.content')" :auto-grow="true" :rules="ruleRequired" required autofocus></v-textarea>
@@ -22,9 +25,10 @@
 <script>
   import { mapGetters } from 'vuex';
   import TagPicker from "./TagPicker";
+  import ColorPicker from "./ColorPicker";
   export default {
     name: "NoteFormText",
-    components: {TagPicker},
+    components: {ColorPicker, TagPicker},
     props: {
       formId: {
         type: String,
@@ -43,11 +47,16 @@
         type: Array,
         required: false,
         default: () => []
+      },
+      initialColor: {
+        type: String,
+        required: false,
       }
     },
     data() {
       let note = {
         title: "",
+        color: this.initialColor,
         tags: this.initialTags,
         content: "",
         markdown: false
@@ -56,6 +65,7 @@
       if(this.data) {
         note.title = this.data.title === this.$t('note.untitled') ? null : this.data.title
         note.tags = this.data.tags
+        note.color = this.data.color
         note.markdown = !!this.data.content.markdown
         note.content = note.markdown ? this.data.content.markdown : this.data.content.text
       }
@@ -91,6 +101,7 @@
         let data = {}
         data.title = this.note.title ? this.note.title : this.$t('note.untitled')
         data.tags = this.note.tags
+        data.color = this.note.color
         data.content = {}
         if(this.note.markdown) data.content.markdown = this.note.content
         else data.content.text = this.note.content
@@ -106,6 +117,7 @@
       data(newData) {
         this.note.title = newData.title === this.$t('note.untitled') ? null : newData.title
         this.note.tags = newData.tags
+        this.note.color = newData.color
         this.note.markdown = !!newData.content.markdown
         this.note.content = this.note.markdown ? newData.content.markdown : newData.content.text
       }